Added shell_commands.php in config.local.yml to override the PHP path.
A new bash script named "cloudy.read_local_config.sh" has been created to read a PHP override in *.local.yml and set it to CLOUDY_PHP. This script incorporates functions that find the configuration file, parse the YAML and return the PHP override. Furthermore, adjustments were made to ensure a proper exit if PHP was not found and to include cloudy.read_local_config.sh in the initial configuration. Also, the version is updated to 1.7.0.